requirements.txt
python-docx>=0.8.11  # optional: only required for --format docx
